% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.sample.aae.custos.R
\name{smp.aae.custo}
\alias{smp.aae.custo}
\title{Stratified Random Sampling Costs}
\usage{
smp.aae.custo(data = NULL, c.lev = 95, d = NULL)
}
\arguments{
\item{data}{A data frame with x and y variables.}

\item{c.lev}{Confidence Level.}

\item{d}{The deviation for calculation of sample size.}
}
\description{
A function that returns the results from calculated estimated values for Stratified Sampling based on Costs.
}
\author{
Marco Aurelio Valles Leal
}
