% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.sample.aee.neyman.R
\name{smp.aae.neyman}
\alias{smp.aae.neyman}
\title{Stratified Random Sampling Neyman}
\usage{
smp.aae.neyman(data = NULL, d = NULL, c.lev = 95)
}
\arguments{
\item{data}{A data frame with x and y variables.}

\item{d}{The deviation for calculation of sample size.}

\item{c.lev}{Confidence Level.}
}
\description{
A function that returns the results from calculated estimated values for Stratified Sampling.
}
\author{
Marco Aurelio Valles Leal
}
