% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.sample.aae.proporcional.R
\name{smp.aae.prop}
\alias{smp.aae.prop}
\title{Stratified Random Sampling Proportion}
\usage{
smp.aae.prop(data = NULL, d = NULL, c.lev = 95)
}
\arguments{
\item{data}{A data frame with x and y variables.}

\item{d}{The deviation for calculation of sample size.}

\item{c.lev}{Confidence Level.}
}
\description{
A function that returns the results from calculated estimated values for Stratified Sampling based on Proportions.
}
\author{
Marco Aurelio Valles Leal
}
