% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.sample.aas.prop.R
\name{smp.aas.prop}
\alias{smp.aas.prop}
\title{Simple Random Sampling for Proportion}
\usage{
smp.aas.prop(c.lev, p = 0.5, dprop = 0.05, population = NULL,
  n = NULL)
}
\arguments{
\item{c.lev}{Confidence Level.}

\item{p}{Estimated proportion.}

\item{dprop}{The deviation for sample size calculation.}

\item{population}{Population size.}

\item{n}{Sample size for non existence of a sample.}
}
\description{
A function that returns the results from calculated estimated values for Stratified Sampling.
}
\author{
Marco Aurelio Valles Leal
}
