% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.sample.aas.R
\name{smp.aas}
\alias{smp.aas}
\title{Simple Random Sampling}
\usage{
smp.aas(x = NULL, c.lev, population = NULL, n = NULL, d = NULL,
  mean = NULL, sd = NULL)
}
\arguments{
\item{x}{A vector containing the data of the sample.}

\item{c.lev}{Confidence Level.}

\item{population}{Population size, for finity population calculation.}

\item{n}{Sample size for non existence of sample data.}

\item{d}{The deviation for sample size calculation.}

\item{mean}{Mean value for non existence of sample data.}

\item{sd}{Standard deviation for non existence of sample data.}
}
\description{
A function that returns the estimated values e confidence values for Proportion.
}
\author{
Marco Aurelio Valles Leal
}
