% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.sample.estim.razao.R
\name{smp.est.razao}
\alias{smp.est.razao}
\title{Ratio Estimators}
\usage{
smp.est.razao(data = NULL, c.lev = 95, N = NULL, mu = NULL)
}
\arguments{
\item{data}{A data frame that contains x and y variables.}

\item{c.lev}{Confidence Level.}

\item{N}{Population size linked to x variable.}

\item{mu}{Population sugested mean.}
}
\description{
A function that returns the results from calculated estimated values for Ratio Estimators.
}
\examples{
x=c(100000,50000,75000,200000,150000,175000,150000)
y=c(300000,200000,300000,600000,450000,520000,450000)
despesas.med=data.frame(area=c(1:7),x=x,y=y)
smp.est.razao(data=despesas.med,N=49)


}
\author{
Marco Aurelio Valles Leal
}
