% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.sample.sistematica.R
\name{smp.sistematica}
\alias{smp.sistematica}
\title{Sistematic Sampling}
\usage{
smp.sistematica(data = NULL, c.lev = 95, N = NULL, type = NULL,
  c.unit = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{c.lev}{Confidence Level.}

\item{N}{Size of population.}

\item{type}{Put "p" for estimate values for proportion.}

\item{c.unit}{TRUE if want to transpose data frame.}
}
\description{
A function that returns the results from calculated estimated values for Sistematic Sampling.
}
\author{
Marco Aurelio Valles Leal
}
